+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/FlagdProvider.java
@@ -3,6 +3,10 @@ package dev.openfeature.contrib.providers.flagd;
 import com.google.protobuf.Descriptors.FieldDescriptor;
 import com.google.protobuf.Message;
 import com.google.protobuf.NullValue;
+import dev.openfeature.contrib.providers.flagd.grpc.EventStreamCallback;
+import dev.openfeature.contrib.providers.flagd.grpc.EventStreamObserver;
+import dev.openfeature.contrib.providers.flagd.grpc.FlagdGrpcInterceptor;
+import dev.openfeature.contrib.providers.flagd.grpc.SslConfigException;
 import dev.openfeature.contrib.providers.flagd.strategy.ResolveStrategy;
 import dev.openfeature.contrib.providers.flagd.strategy.SimpleResolving;
 import dev.openfeature.contrib.providers.flagd.strategy.TracedResolving;
