--- providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/Protocol.java
@@ -1,9 +1,0 @@
-package dev.openfeature.contrib.providers.flagd;
-
-/**
- * Protocol for gRPC connect.
- */
-public enum Protocol {
-    HTTP,
-    HTTPS
-}
