--- providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/SslConfigException.java
@@ -1,7 +1,0 @@
-package dev.openfeature.contrib.providers.flagd;
-
-class SslConfigException extends RuntimeException {
-    public SslConfigException(String message) {
-        super(message);
-    }
-}
