+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/grpc/FlagdGrpcInterceptor.java
@@ -1,4 +1,4 @@
-package dev.openfeature.contrib.providers.flagd;
+package dev.openfeature.contrib.providers.flagd.grpc;
 
 import io.grpc.CallOptions;
 import io.grpc.Channel;
@@ -18,12 +18,12 @@ import javax.annotation.Nullable;
  * FlagdGrpcInterceptor is an interceptor for grpc communication from java-sdk to flagd.
  * <a href="https://github.com/open-telemetry/opentelemetry-java-docs">credits</a>
  */
-final class FlagdGrpcInterceptor implements ClientInterceptor {
+final public class FlagdGrpcInterceptor implements ClientInterceptor {
     private static final TextMapSetter<Metadata> SETTER = new Setter();
 
     private final OpenTelemetry openTelemetry;
 
-    FlagdGrpcInterceptor(@Nonnull final OpenTelemetry openTelemetry) {
+    public FlagdGrpcInterceptor(@Nonnull final OpenTelemetry openTelemetry) {
         this.openTelemetry = openTelemetry;
     }
