+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/grpc/SslConfigException.java
@@ -1,0 +1,7 @@
+package dev.openfeature.contrib.providers.flagd.grpc;
+
+public class SslConfigException extends RuntimeException {
+    public SslConfigException(String message) {
+        super(message);
+    }
+}
